bug(events): bugfix failing Intercom.Events.submit/1
Within the last changes, a bug was introduced by passing the field
`user_id` instead of `id` to the /events endpoint. This means that if a
intercom_id is passed as it is described in the docs (and orgininally
intended), the request fails because the intercom_id is passed as the
user_id. This commit reverts the change made and aligns the code to the
documentation.